--- controllers/auth.js
@@ -1,27 +1,0 @@
-const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
-
-const { User } = require("../models/user");
-
-const { HttpError, ctrlWrapper } = require("../helpers");
-
-const register = async (req, res) => {
-  const { email } = req.body;
-  const user = await User.findOne({ email });
-
-  if (user) {
-    throw HttpError(409, "Email in use");
-  }
-
-  const newUser = await User.create(req.body);
-  res.status(201).json({
-    user: {
-      email: newUser.email,
-      subscription: newUser.subscription,
-    },
-  });
-};
-
-module.exports = {
-  register: ctrlWrapper(register),
-};
